gitの補完設定で初回のみ必要な動作を繰り返し実行しないように修正

+++ .zshrc
@@ -42,9 +42,16 @@ export PATH=/Users/3no3/Library/Android/sdk/platform-tools:$PATH
 # gitコマンドの補完設定
 mkdir -p ~/.zsh/completion
 cd ~/.zsh/completion
-curl -O https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
-curl -O https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.zsh
-mv git-completion.zsh _git
+
+if [[ ! -f ./git-completion.bash ]]; then
+  curl -O https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
+fi
+
+if [[ ! -f ./_git ]]; then
+  curl -O https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.zsh
+  mv git-completion.zsh _git
+fi
+
 fpath=(~/.zsh/completion $fpath)
 cd -
